// types/camper.ts
export interface CamperImage {
  original: string;
  thumb: string;
}

export interface CamperReview {
  reviewer_name: string;
  reviewer_rating: number;
  comment: string;
}

export interface Camper {
  id: string;
  name: string;
  price: number;
  rating: number;
  location: string;
  description: string;

  transmission?: string;
  engine?: string;
  AC?: boolean;
  bathroom?: boolean;
  kitchen?: boolean;
  TV?: boolean;
  radio?: boolean;
  refrigerator?: boolean;
  microwave?: boolean;
  gas?: boolean;
  water?: boolean;

  form?: string;
  length?: string;
  width?: string;
  height?: string;
  tank?: string;
  consumption?: string;

  gallery: CamperImage[];
  reviews: CamperReview[];
}